slack_gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
google-auth==2.34.0
requests>=2.31.0
